Optimization: query objects as subclasses
The `ModelWithInheritance.as_leaf_class` method, which is widely used in
A+ to fetch polymorphic objects as their actual subclasses, is very bad
for performance because it executes a new database query every time it
is called. In this commit, InheritanceManager from django-model-utils is
used to avoid calling `as_leaf_class` whenever possible.
InheritanceManager adds a new `select_subclasses` method to querysets,
which returns objects as their actual subclasses without extra queries.